core/testing/agents/synthetic_data/synthetic_data_generator_agent_specific_mapped/requirements.txt
anthropic>=0.39.0
pyyaml>=6.0.0
claude-agent-sdk
anyio
